# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "https://archimed.justcode.am/",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/doctors_list": {
            "post": {
                "tags": [
                    "Doctors"
                ],
                "summary": "doctors_list",
                "description": "Получаем Список врачей",
                "operationId": "doctors_list",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "city": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Москва"
                                    },
                                    "orderby": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Doctors Data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/single_doctor": {
            "post": {
                "tags": [
                    "Doctors"
                ],
                "summary": "single_doctor",
                "description": "Получаем Одного врача",
                "operationId": "single_doctor",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "doctor_id": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Doctor Data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/doctor_visit_time": {
            "post": {
                "tags": [
                    "Doctors"
                ],
                "summary": "doctor_visit_time",
                "description": "Получаем Список свободных часов Одного доктора для конкретной даты",
                "operationId": "doctor_visit_time",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "subject_id": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "PL_EXAM_ID": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2023-05-18"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2023-05-18"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Visit Time Data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong Data sended subject_id and PL_EXAM_ID no exist",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/api/get_auth_user_order": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "get auth user orders",
                "description": "get_auth_user_order",
                "operationId": "get_auth_user_order",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Basket data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}